+++ BUILD.bazel
@@ -1613,10 +1613,10 @@ cc_test(
     copts = COPTS,
     tags = ["team:core"],
     deps = [
-        ":ray_mock",
         ":gcs_server_lib",
         ":gcs_server_test_util",
         ":gcs_test_util_lib",
+        ":ray_mock",
         "@com_google_googletest//:gtest_main",
     ],
 )
@@ -1713,6 +1713,7 @@ cc_test(
         ":gcs_server_lib",
         ":gcs_server_test_util",
         ":gcs_test_util_lib",
+        ":ray_mock",
         "@com_google_googletest//:gtest_main",
     ],
 )
