+++ python/ray/tests/test_basic_5.py
@@ -6,11 +6,8 @@ import sys
 import time
 import pytest
 
-import psutil
 import ray.cluster_utils
-from ray._private.test_utils import client_test_enabled, wait_for_condition
-from ray._private.test_utils import wait_for_pid_to_exit
-from ray.experimental.internal_kv import _internal_kv_list
+from ray._private.test_utils import wait_for_pid_to_exit, client_test_enabled
 
 import ray
 
@@ -119,101 +116,6 @@ def test_internal_kv(ray_start_regular):
         kv._internal_kv_del("@namespace_def", namespace="n")
     with pytest.raises(RuntimeError):
         kv._internal_kv_list("@namespace_abc", namespace="n")
-
-
-def get_gcs_memory_used():
-    m = sum([
-        process.memory_info().rss for process in psutil.process_iter()
-        if process.name() in ("gcs_server", "redis-server")
-    ])
-    return m
-
-
-def function_entry_num(job_id):
-    from ray.ray_constants import KV_NAMESPACE_FUNCTION_TABLE
-    return len(_internal_kv_list(b"IsolatedExports:" + job_id,
-                                 namespace=KV_NAMESPACE_FUNCTION_TABLE)) + \
-        len(_internal_kv_list(b"RemoteFunction:" + job_id,
-                              namespace=KV_NAMESPACE_FUNCTION_TABLE)) + \
-        len(_internal_kv_list(b"ActorClass:" + job_id,
-                              namespace=KV_NAMESPACE_FUNCTION_TABLE))
-
-
-@pytest.mark.skipif(
-    client_test_enabled(),
-    reason="client api doesn't support namespace right now.")
-def test_function_table_gc(call_ray_start):
-    """This test tries to verify that function table is cleaned up
-    after job exits.
-    """
-
-    def f():
-        data = "0" * 1024 * 1024  # 1MB
-
-        @ray.remote
-        def r():
-            nonlocal data
-
-            @ray.remote
-            class Actor:
-                pass
-
-        return r.remote()
-
-    ray.init(address="auto", namespace="b")
-
-    # It should use > 500MB data
-    ray.get([f() for _ in range(500)])
-
-    assert get_gcs_memory_used() > 500 * 1024 * 1024
-    job_id = ray.worker.global_worker.current_job_id.binary()
-    assert function_entry_num(job_id) > 0
-    ray.shutdown()
-
-    # We only check it for redis mode since in memory mode, the rss depends
-    # on when the OS evict the page.
-    if os.environ.get("RAY_gcs_storage") != "memory":
-        wait_for_condition(
-            lambda: get_gcs_memory_used() < 100 * 1024 * 1024, timeout=30)
-
-    # now check the function table is cleaned up after job finished
-    ray.init(address="auto", namespace="a")
-    wait_for_condition(lambda: function_entry_num(job_id) == 0)
-
-
-@pytest.mark.skipif(
-    client_test_enabled(),
-    reason="client api doesn't support namespace right now.")
-def test_function_table_gc_actor(call_ray_start):
-    """If there is a detached actor, the table won't be cleaned up.
-    """
-    ray.init(address="auto", namespace="a")
-
-    @ray.remote
-    class Actor:
-        def ready(self):
-            return
-
-    # If there is a detached actor, the function won't be deleted.
-    a = Actor.options(lifetime="detached", name="a").remote()
-    ray.get(a.ready.remote())
-    job_id = ray.worker.global_worker.current_job_id.binary()
-    ray.shutdown()
-
-    ray.init(address="auto", namespace="b")
-    with pytest.raises(Exception):
-        wait_for_condition(lambda: function_entry_num(job_id) == 0)
-    a = ray.get_actor("a", namespace="a")
-    ray.kill(a)
-    wait_for_condition(lambda: function_entry_num(job_id) == 0)
-
-    # If there is not a detached actor, it'll be deleted when the job finishes.
-    a = Actor.remote()
-    ray.get(a.ready.remote())
-    job_id = ray.worker.global_worker.current_job_id.binary()
-    ray.shutdown()
-    ray.init(address="auto", namespace="c")
-    wait_for_condition(lambda: function_entry_num(job_id) == 0)
 
 
 if __name__ == "__main__":
