+++ src/ray/gcs/gcs_server/test/gcs_actor_manager_test.cc
@@ -14,11 +14,15 @@
 
 #include <memory>
 
+// clang-format off
 #include "gtest/gtest.h"
 #include "ray/common/asio/instrumented_io_context.h"
 #include "ray/common/test_util.h"
 #include "ray/gcs/gcs_server/test/gcs_server_test_util.h"
 #include "ray/gcs/test/gcs_test_util.h"
+#include "ray/gcs/gcs_server/gcs_kv_manager.h"
+#include "mock/ray/gcs/gcs_server/gcs_kv_manager.h"
+// clang-format on
 
 namespace ray {
 
@@ -107,9 +111,11 @@ class GcsActorManagerTest : public ::testing::Test {
         std::make_unique<GcsServerMocker::MockGcsPubSub>(redis_client_));
     store_client_ = std::make_shared<gcs::InMemoryStoreClient>(io_service_);
     gcs_table_storage_ = std::make_shared<gcs::InMemoryGcsTableStorage>(io_service_);
-    gcs_actor_manager_.reset(new gcs::GcsActorManager(
+    kv_ = std::make_unique<gcs::MockInternalKVInterface>();
+    function_manager_ = std::make_unique<gcs::GcsFunctionManager>(*kv_);
+    gcs_actor_manager_ = std::make_unique<gcs::GcsActorManager>(
         io_service_, mock_actor_scheduler_, gcs_table_storage_, gcs_publisher_,
-        *runtime_env_mgr_, [](const ActorID &actor_id) {},
+        *runtime_env_mgr_, *function_manager_, [](const ActorID &actor_id) {},
         [this](const JobID &job_id) {
           auto job_config = std::make_shared<rpc::JobConfig>();
           job_config->set_ray_namespace(job_namespace_table_[job_id]);
@@ -123,7 +129,7 @@ class GcsActorManagerTest : public ::testing::Test {
             delayed_to_run_ = fn;
           }
         },
-        [this](const rpc::Address &addr) { return worker_client_; }));
+        [this](const rpc::Address &addr) { return worker_client_; });
 
     for (int i = 1; i <= 10; i++) {
       auto job_id = JobID::FromInt(i);
@@ -218,6 +224,8 @@ class GcsActorManagerTest : public ::testing::Test {
   const std::chrono::milliseconds timeout_ms_{2000};
   std::function<void(void)> delayed_to_run_;
   boost::posix_time::milliseconds delay_;
+  std::unique_ptr<gcs::GcsFunctionManager> function_manager_;
+  std::unique_ptr<gcs::MockInternalKVInterface> kv_;
   bool skip_delay_;
 };
